+++ example/lib/main.dart
@@ -938,14 +938,27 @@ class MyHomePageState extends State<MyHomePage> {
                           ChoiceChipField<int>(
                         label: 'Frutas',
                         enabled: edit,
-                        items: const <int, String>{
-                          0: 'Banana',
-                          1: 'Maça',
-                          2: 'Laranja',
+                        items: const <int, ChipEntry>{
+                          0: ChipEntry(
+                            '🍎Maça',
+                            color: Colors.red,
+                            selectedColor: Colors.redAccent,
+                          ),
+                          1: ChipEntry(
+                            '🍌Banana',
+                            color: Colors.yellow,
+                            selectedColor: Colors.yellowAccent,
+                          ),
+                          2: ChipEntry(
+                            '🍊Tangerina',
+                            color: Colors.orange,
+                            selectedColor: Colors.orangeAccent,
+                          ),
                         },
-                        onChanged: (int? value) =>
+                        onChanged: (int? value, {required bool selected}) =>
                             // ignore: avoid_print
-                            print('ChoiceChipField changed to $value'),
+                            print('ChoiceChipField $value is'
+                                '${selected ? '' : ' NOT'} selected'),
                         validator: FollyValidators.notNull,
                         onSaved: (int? value) => model.fruitIndex = value,
                       ),