+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Updating to Font Awesome 10.7.0.
 * Updating analysis_options.yaml.
-* 
+* Updating ChoiceChipField style.
 
 ## [2.6.1] - 2024-01-27